Deploy any Azure Template with Parameters
this script will run any azure bicep or ARM template, read in a .json parameters file, and prompt the user to change any of the available parameters at runtime. Any params the user doesn't enter a value for will default to the value entered in the parameters file. By default, the script will not prompt the user for any password-related parameters at runtime.